Lesson 5-1 working code

+++ src/services/user.service.ts
@@ -12,23 +12,35 @@ class UserService {
         return userRepository.create(user);
     }
 
-    public getById(userId: string): Promise<IUser> {
-        return userRepository.getById(userId);
+    public async getById(userId: string): Promise<IUser> {
+        const user = await userRepository.getById(userId);
+        if (!user) {
+            throw new ApiError("User not found", StatusCodeEnum.NOT_FOUND);
+        }
+        return user;
     }
 
-    public updateById(userId: string, user: IUserUpdateDTO): Promise<IUser> {
-        return userRepository.updateById(userId, user);
+    public async updateById(userId: string, user: IUserUpdateDTO): Promise<IUser> {
+        const data = await userRepository.getById(userId);
+        if (!data) {
+            throw new ApiError("User not found", StatusCodeEnum.NOT_FOUND);
+        }
+        return await userRepository.updateById(userId, user);
     }
 
-    public deleteById(userId: string): Promise<IUser> {
-        return userRepository.deleteById(userId);
+    public async deleteById(userId: string): Promise<IUser> {
+        const data = await userRepository.getById(userId);
+        if (!data) {
+            throw new ApiError("User not found", StatusCodeEnum.NOT_FOUND);
+        }
+        return await userRepository.deleteById(userId);
     }
 
     public async isEmailUnique(email: string): Promise<void> {
-    const user = await userRepository.getByEmail(email);
-    if (user) {
-        throw new ApiError("User is already exist", StatusCodeEnum.BED_REQUEST)
-    }
+        const user = await userRepository.getByEmail(email);
+        if (user) {
+            throw new ApiError("User is already exist", StatusCodeEnum.BED_REQUEST);
+        }
     }
 }
